requirements.txt
pydantic>=2.0.0
scikit-learn>=1.2.0
pandas>=2.0.0
numpy>=1.22.0
matplotlib>=3.5.0
seaborn>=0.12.0
flaml>=2.0.0
PyYAML>=6.0
joblib>=1.1.0